python/examples/network_compute_bridge/docker-requirements_client.txt
bosdyn-api==3.2.0
bosdyn-client==3.2.0
bosdyn-core==3.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
Deprecated==1.2.13
grpcio==1.42.0
idna==3.3
numpy==1.19.4
opencv-python==4.5.1.48
protobuf==3.19.1
PyJWT==2.3.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
wrapt==1.13.3
